+++ src/Shapes.jl
@@ -184,8 +184,13 @@ Append a new shape to this world
 
 See also: [`Shape`](@ref), [`World`](@ref)
 """
-function add_shape!(W::World, S::Shape)
-    push!(W.shapes, S)
+function add_shape!(world::World, S::Shape)
+    push!(world.shapes, S)
+    return nothing
+end
+
+function add_light!(world::World, PL::PointLight)
+    push!(world.point_lights, PL)
     return nothing
 end
 
@@ -210,4 +215,84 @@ function ray_intersection(world::World, ray::Ray)
     end
     
     return closest
+end
+
+##########################################################################################92
+
+"""
+    quick_ray_intersection(shape::Shape, ray::Ray) :: Bool
+
+Quickly determine whether a ray hits the shape or not.
+
+See also: [`Shape`](@ref), [`Ray`](@ref)
+"""
+function quick_ray_intersection(shape::Shape, ray::Ray)
+    return ErrorException(
+            "quick_ray_intersection is an abstract method"*
+            "and cannot be called directlly")
+end
+
+"""
+    quick_ray_intersection(sphere::Sphere, ray::Ray) :: Bool
+
+Quickly checks if the `ray` intersects the `sphere`.
+
+See also: [`Sphere`](@ref), [`Ray`](@ref)
+"""
+function quick_ray_intersection(sphere::Sphere, ray::Ray)
+    inv_ray = inverse(sphere.T) * ray
+    origin_vec = Vec(inv_ray.origin)
+
+    a = squared_norm(inv_ray.dir)
+    b = 2.0 * origin_vec ⋅ inv_ray.dir
+    c = squared_norm(origin_vec) - 1.0
+    Δ = b * b - 4.0 * a * c 
+     
+    (Δ > 0.0) || (return false)
+
+    tmin = (-b - √Δ) / (2.0 * a)
+    tmax = (-b + √Δ) / (2.0 * a)
+
+    return ((inv_ray.tmin < tmin < inv_ray.tmax) || (inv_ray.tmin < tmax < inv_ray.tmax))
+end
+
+"""
+    quick_ray_intersection(plane::Plane, ray::Ray) :: Bool
+
+Quickly checks if the `ray` intersects the `plane`.
+
+See also: [`Plane`](@ref), [`Ray`](@ref)
+"""
+function quick_ray_intersection(plane::Plane, ray::Ray)
+    inv_ray = inverse(plane.T) * ray
+    !(inv_ray.dir.z ≈ 0.) || (return false)
+
+    t = -inv_ray.origin.z / inv_ray.dir.z
+    return (inv_ray.tmin < t < inv_ray.tmax)
+end
+
+
+"""
+    is_point_visible(
+        world::World, 
+        point::Point, 
+        observer_pos::Point
+        ) :: Bool
+
+Return `true` if the straight line connecting `observer_pos` to `point`
+do not intersect any of the shapes of `world` between the two points,
+otherwise return `false`.
+
+See also: [`World`](@ref), [`Point`](@ref)
+"""
+function is_point_visible(world::World, point::Point, observer_pos::Point)
+    direction = point - observer_pos
+    dir_norm = norm(direction)
+
+    ray = Ray(observer_pos, direction, 1e-2 / dir_norm, 1.0)
+    for shape in world.shapes
+        (quick_ray_intersection(shape, ray) == false) || (return false)
+    end
+
+    return true
 end